+++ webapp/javascript/pages/ContinuousDiffView.tsx
@@ -26,7 +26,7 @@ import SyncTimelines from '@webapp/components/TimelineChart/SyncTimelines';
 import useExportToFlamegraphDotCom from '@webapp/components/exportToFlamegraphDotCom.hook';
 import { LoadingOverlay } from '@webapp/ui/LoadingOverlay';
 import ExportData from '@webapp/components/ExportData';
-import TimelineTitle from '@webapp/components/TimelineTitle';
+import ChartTitle from '@webapp/components/ChartTitle';
 import { isExportToFlamegraphDotComEnabled } from '@webapp/util/features';
 import PageTitle from '@webapp/components/PageTitle';
 import { formatTitle } from './formatTitle';
@@ -140,9 +140,7 @@ function ComparisonDiffApp() {
               }}
               selectionType="double"
               timezone={timezone}
-              title={
-                <TimelineTitle titleKey={diffView.profile?.metadata.units} />
-              }
+              title={<ChartTitle titleKey={diffView.profile?.metadata.units} />}
             />
             <SyncTimelines
               timeline={leftTimeline}
@@ -157,7 +155,7 @@ function ComparisonDiffApp() {
         <div className="diff-instructions-wrapper">
           <Box className="diff-instructions-wrapper-side">
             <LoadingOverlay active={isLoading}>
-              <TimelineTitle titleKey="baseline" color={leftColor} />
+              <ChartTitle titleKey="baseline" color={leftColor} />
               <TagsBar
                 query={leftQuery}
                 tags={leftTags}
@@ -191,7 +189,7 @@ function ComparisonDiffApp() {
           </Box>
           <Box className="diff-instructions-wrapper-side">
             <LoadingOverlay active={isLoading}>
-              <TimelineTitle titleKey="comparison" color={rightColor} />
+              <ChartTitle titleKey="comparison" color={rightColor} />
               <TagsBar
                 query={rightQuery}
                 tags={rightTags}
@@ -226,7 +224,7 @@ function ComparisonDiffApp() {
         </div>
         <Box>
           <LoadingOverlay active={isLoading} spinnerPosition="baseline">
-            <TimelineTitle titleKey="diff" />
+            <ChartTitle titleKey="diff" />
             <FlamegraphRenderer
               showCredit={false}
               profile={diffView.profile}