+++ webapp/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.57.0](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.9...@pyroscope/webapp@1.57.0) (2022-11-09)
+
+
+### Features
+
+* add single, comparison, diff tabs to heatmap page ([#1672](https://github.com/pyroscope-io/pyroscope/issues/1672)) ([9afe5e5](https://github.com/pyroscope-io/pyroscope/commit/9afe5e564e1fea6fe4026c3180134412ac2930f9))
+
+
+
+
+
+## [1.56.9](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.8...@pyroscope/webapp@1.56.9) (2022-11-09)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.56.8](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.7...@pyroscope/webapp@1.56.8) (2022-11-09)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.56.7](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.6...@pyroscope/webapp@1.56.7) (2022-11-09)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
+## [1.56.6](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.5...@pyroscope/webapp@1.56.6) (2022-11-08)
+
+**Note:** Version bump only for package @pyroscope/webapp
+
+
+
+
+
 ## [1.56.5](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/webapp@1.56.4...@pyroscope/webapp@1.56.5) (2022-11-08)